+++ tests/scanner_assets/client_code.rs
@@ -22,6 +22,7 @@ pub mod wl_foo {
     #[doc = "Possible cake kinds\n\nList of the possible kind of cake supported by the protocol."]
     #[repr(u32)]
     #[derive(Copy, Clone, Debug, PartialEq)]
+    #[non_exhaustive]
     pub enum CakeKind {
         #[doc = "mild cake without much flavor"]
         Basic = 0,
@@ -29,8 +30,6 @@ pub mod wl_foo {
         Spicy = 1,
         #[doc = "fruity cake to get vitamins"]
         Fruity = 2,
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl CakeKind {
         pub fn from_raw(n: u32) -> Option<CakeKind> {
@@ -54,6 +53,7 @@ pub mod wl_foo {
             self.bits()
         }
     }
+    #[non_exhaustive]
     pub enum Request {
         #[doc = "do some foo\n\nThis will do some foo with its args."]
         FooIt {
@@ -65,8 +65,6 @@ pub mod wl_foo {
         },
         #[doc = "create a bar\n\nCreate a bar which will do its bar job."]
         CreateBar {},
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[
@@ -92,20 +90,17 @@ pub mod wl_foo {
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::FooIt { .. } => 0,
                 Request::CreateBar { .. } => 1,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::FooIt { .. } => 1,
                 Request::CreateBar { .. } => 1,
             }
@@ -124,7 +119,6 @@ pub mod wl_foo {
         }
         fn into_raw(self, sender_id: u32) -> Message {
             match self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::FooIt {
                     number,
                     unumber,
@@ -163,7 +157,6 @@ pub mod wl_foo {
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
             match self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::FooIt {
                     number,
                     unumber,
@@ -188,11 +181,10 @@ pub mod wl_foo {
             }
         }
     }
+    #[non_exhaustive]
     pub enum Event {
         #[doc = "a cake is possible\n\nThe server advertises that a kind of cake is available\n\nOnly available since version 2 of the interface"]
         Cake { kind: CakeKind, amount: u32 },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[super::MessageDesc {
@@ -204,19 +196,16 @@ pub mod wl_foo {
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Cake { .. } => 0,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Cake { .. } => 2,
             }
         }
@@ -375,6 +364,7 @@ pub mod wl_bar {
         MessageGroup, Object, ObjectMetadata, Proxy, NULLPTR,
     };
     use std::os::raw::c_char;
+    #[non_exhaustive]
     pub enum Request {
         #[doc = "ask for a bar delivery\n\nProceed to a bar delivery of given foo.\n\nOnly available since version 2 of the interface"]
         BarDelivery {
@@ -396,8 +386,6 @@ pub mod wl_bar {
             request: u32,
             event: u32,
         },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[
@@ -437,14 +425,12 @@ pub mod wl_bar {
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::Release => true,
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::BarDelivery { .. } => 0,
                 Request::Release => 1,
                 Request::_Self { .. } => 2,
@@ -452,7 +438,6 @@ pub mod wl_bar {
         }
         fn since(&self) -> u32 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::BarDelivery { .. } => 2,
                 Request::Release => 1,
                 Request::_Self { .. } => 2,
@@ -468,7 +453,6 @@ pub mod wl_bar {
         }
         fn into_raw(self, sender_id: u32) -> Message {
             match self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::BarDelivery {
                     kind,
                     target,
@@ -526,7 +510,6 @@ pub mod wl_bar {
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
             match self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::BarDelivery {
                     kind,
                     target,
@@ -581,6 +564,7 @@ pub mod wl_bar {
             }
         }
     }
+    #[non_exhaustive]
     pub enum Event {
         #[doc = "ask for erronous bindings from wayland-scanner\n\nThis event tests argument names which can break wayland-scanner.\n\nOnly available since version 2 of the interface"]
         _Self {
@@ -593,8 +577,6 @@ pub mod wl_bar {
             request: u32,
             event: u32,
         },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[super::MessageDesc {
@@ -615,19 +597,16 @@ pub mod wl_bar {
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::_Self { .. } => 0,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::_Self { .. } => 2,
             }
         }
@@ -867,27 +846,19 @@ pub mod wl_display {
         MessageGroup, Object, ObjectMetadata, Proxy, NULLPTR,
     };
     use std::os::raw::c_char;
-    pub enum Request {
-        #[doc(hidden)]
-        __nonexhaustive,
-    }
+    #[non_exhaustive]
+    pub enum Request {}
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[];
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn opcode(&self) -> u16 {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn since(&self) -> u32 {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
             match opcode {
@@ -898,9 +869,7 @@ pub mod wl_display {
             panic!("Request::from_raw can not be used Client-side.")
         }
         fn into_raw(self, sender_id: u32) -> Message {
-            match self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match self {}
         }
         unsafe fn from_raw_c(
             obj: *mut ::std::os::raw::c_void,
@@ -913,32 +882,22 @@ pub mod wl_display {
         where
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
-            match self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match self {}
         }
     }
-    pub enum Event {
-        #[doc(hidden)]
-        __nonexhaustive,
-    }
+    #[non_exhaustive]
+    pub enum Event {}
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[];
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
-            match *self {
-                Event::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn opcode(&self) -> u16 {
-            match *self {
-                Event::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn since(&self) -> u32 {
-            match *self {
-                Event::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
             match opcode {
@@ -1018,11 +977,10 @@ pub mod wl_registry {
         MessageGroup, Object, ObjectMetadata, Proxy, NULLPTR,
     };
     use std::os::raw::c_char;
+    #[non_exhaustive]
     pub enum Request {
         #[doc = "bind an object to the display\n\nThis request is a special code-path, as its new-id argument as no target type."]
         Bind { name: u32, id: (String, u32) },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[super::MessageDesc {
@@ -1034,19 +992,16 @@ pub mod wl_registry {
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::Bind { .. } => 0,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::Bind { .. } => 1,
             }
         }
@@ -1060,7 +1015,6 @@ pub mod wl_registry {
         }
         fn into_raw(self, sender_id: u32) -> Message {
             match self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::Bind { name, id } => Message {
                     sender_id: sender_id,
                     opcode: 0,
@@ -1087,7 +1041,6 @@ pub mod wl_registry {
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
             match self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::Bind { name, id } => {
                     let mut _args_array: [wl_argument; 4] = unsafe { ::std::mem::zeroed() };
                     _args_array[0].u = name;
@@ -1100,27 +1053,19 @@ pub mod wl_registry {
             }
         }
     }
-    pub enum Event {
-        #[doc(hidden)]
-        __nonexhaustive,
-    }
+    #[non_exhaustive]
+    pub enum Event {}
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[];
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
-            match *self {
-                Event::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn opcode(&self) -> u16 {
-            match *self {
-                Event::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn since(&self) -> u32 {
-            match *self {
-                Event::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
             match opcode {
@@ -1221,27 +1166,19 @@ pub mod wl_callback {
         MessageGroup, Object, ObjectMetadata, Proxy, NULLPTR,
     };
     use std::os::raw::c_char;
-    pub enum Request {
-        #[doc(hidden)]
-        __nonexhaustive,
-    }
+    #[non_exhaustive]
+    pub enum Request {}
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[];
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn opcode(&self) -> u16 {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn since(&self) -> u32 {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
             match opcode {
@@ -1252,9 +1189,7 @@ pub mod wl_callback {
             panic!("Request::from_raw can not be used Client-side.")
         }
         fn into_raw(self, sender_id: u32) -> Message {
-            match self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match self {}
         }
         unsafe fn from_raw_c(
             obj: *mut ::std::os::raw::c_void,
@@ -1267,16 +1202,13 @@ pub mod wl_callback {
         where
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
-            match self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match self {}
         }
     }
+    #[non_exhaustive]
     pub enum Event {
         #[doc = "done event\n\nThis event is actually a destructor, but the protocol XML has no way of specifying it.\nAs such, the scanner should consider wl_callback.done as a special case.\n\nThis is a destructor, once received this object cannot be used any longer."]
         Done { callback_data: u32 },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[super::MessageDesc {
@@ -1288,19 +1220,16 @@ pub mod wl_callback {
         type Map = super::ProxyMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { .. } => true,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { .. } => 0,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { .. } => 1,
             }
         }