+++ tests/scanner_assets/server_code.rs
@@ -22,6 +22,7 @@ pub mod wl_foo {
     #[doc = "Possible cake kinds\n\nList of the possible kind of cake supported by the protocol."]
     #[repr(u32)]
     #[derive(Copy, Clone, Debug, PartialEq)]
+    #[non_exhaustive]
     pub enum CakeKind {
         #[doc = "mild cake without much flavor"]
         Basic = 0,
@@ -29,8 +30,6 @@ pub mod wl_foo {
         Spicy = 1,
         #[doc = "fruity cake to get vitamins"]
         Fruity = 2,
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl CakeKind {
         pub fn from_raw(n: u32) -> Option<CakeKind> {
@@ -54,6 +53,7 @@ pub mod wl_foo {
             self.bits()
         }
     }
+    #[non_exhaustive]
     pub enum Request {
         #[doc = "do some foo\n\nThis will do some foo with its args."]
         FooIt {
@@ -65,8 +65,6 @@ pub mod wl_foo {
         },
         #[doc = "create a bar\n\nCreate a bar which will do its bar job."]
         CreateBar { id: Main<super::wl_bar::WlBar> },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[
@@ -92,20 +90,17 @@ pub mod wl_foo {
         type Map = super::ResourceMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::FooIt { .. } => 0,
                 Request::CreateBar { .. } => 1,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::FooIt { .. } => 1,
                 Request::CreateBar { .. } => 1,
             }
@@ -218,11 +213,10 @@ pub mod wl_foo {
             panic!("Request::as_raw_c_in can not be used Server-side.")
         }
     }
+    #[non_exhaustive]
     pub enum Event {
         #[doc = "a cake is possible\n\nThe server advertises that a kind of cake is available\n\nOnly available since version 2 of the interface"]
         Cake { kind: CakeKind, amount: u32 },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[super::MessageDesc {
@@ -234,19 +228,16 @@ pub mod wl_foo {
         type Map = super::ResourceMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Cake { .. } => 0,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Cake { .. } => 2,
             }
         }
@@ -260,7 +251,6 @@ pub mod wl_foo {
         }
         fn into_raw(self, sender_id: u32) -> Message {
             match self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Cake { kind, amount } => Message {
                     sender_id: sender_id,
                     opcode: 0,
@@ -280,7 +270,6 @@ pub mod wl_foo {
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
             match self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Cake { kind, amount } => {
                     let mut _args_array: [wl_argument; 2] = unsafe { ::std::mem::zeroed() };
                     _args_array[0].u = kind.to_raw();
@@ -375,6 +364,7 @@ pub mod wl_bar {
         MessageGroup, Object, ObjectMetadata, Resource, NULLPTR,
     };
     use std::os::raw::c_char;
+    #[non_exhaustive]
     pub enum Request {
         #[doc = "ask for a bar delivery\n\nProceed to a bar delivery of given foo.\n\nOnly available since version 2 of the interface"]
         BarDelivery {
@@ -396,8 +386,6 @@ pub mod wl_bar {
             request: u32,
             event: u32,
         },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[
@@ -437,14 +425,12 @@ pub mod wl_bar {
         type Map = super::ResourceMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::Release => true,
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::BarDelivery { .. } => 0,
                 Request::Release => 1,
                 Request::_Self { .. } => 2,
@@ -452,7 +438,6 @@ pub mod wl_bar {
         }
         fn since(&self) -> u32 {
             match *self {
-                Request::__nonexhaustive => unreachable!(),
                 Request::BarDelivery { .. } => 2,
                 Request::Release => 1,
                 Request::_Self { .. } => 2,
@@ -619,6 +604,7 @@ pub mod wl_bar {
             panic!("Request::as_raw_c_in can not be used Server-side.")
         }
     }
+    #[non_exhaustive]
     pub enum Event {
         #[doc = "ask for erronous bindings from wayland-scanner\n\nThis event tests argument names which can break wayland-scanner.\n\nOnly available since version 2 of the interface"]
         _Self {
@@ -631,8 +617,6 @@ pub mod wl_bar {
             request: u32,
             event: u32,
         },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[super::MessageDesc {
@@ -653,19 +637,16 @@ pub mod wl_bar {
         type Map = super::ResourceMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 _ => false,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::_Self { .. } => 0,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::_Self { .. } => 2,
             }
         }
@@ -679,7 +660,6 @@ pub mod wl_bar {
         }
         fn into_raw(self, sender_id: u32) -> Message {
             match self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::_Self {
                     _self,
                     _mut,
@@ -717,7 +697,6 @@ pub mod wl_bar {
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
             match self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::_Self {
                     _self,
                     _mut,
@@ -854,27 +833,19 @@ pub mod wl_callback {
         MessageGroup, Object, ObjectMetadata, Resource, NULLPTR,
     };
     use std::os::raw::c_char;
-    pub enum Request {
-        #[doc(hidden)]
-        __nonexhaustive,
-    }
+    #[non_exhaustive]
+    pub enum Request {}
     impl super::MessageGroup for Request {
         const MESSAGES: &'static [super::MessageDesc] = &[];
         type Map = super::ResourceMap;
         fn is_destructor(&self) -> bool {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn opcode(&self) -> u16 {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn since(&self) -> u32 {
-            match *self {
-                Request::__nonexhaustive => unreachable!(),
-            }
+            match *self {}
         }
         fn child<Meta: ObjectMetadata>(opcode: u16, version: u32, meta: &Meta) -> Option<Object<Meta>> {
             match opcode {
@@ -905,11 +876,10 @@ pub mod wl_callback {
             panic!("Request::as_raw_c_in can not be used Server-side.")
         }
     }
+    #[non_exhaustive]
     pub enum Event {
         #[doc = "done event\n\nThis event is actually a destructor, but the protocol XML has no way of specifying it.\nAs such, the scanner should consider wl_callback.done as a special case.\n\nThis is a destructor, once sent this object cannot be used any longer."]
         Done { callback_data: u32 },
-        #[doc(hidden)]
-        __nonexhaustive,
     }
     impl super::MessageGroup for Event {
         const MESSAGES: &'static [super::MessageDesc] = &[super::MessageDesc {
@@ -921,19 +891,16 @@ pub mod wl_callback {
         type Map = super::ResourceMap;
         fn is_destructor(&self) -> bool {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { .. } => true,
             }
         }
         fn opcode(&self) -> u16 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { .. } => 0,
             }
         }
         fn since(&self) -> u32 {
             match *self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { .. } => 1,
             }
         }
@@ -947,7 +914,6 @@ pub mod wl_callback {
         }
         fn into_raw(self, sender_id: u32) -> Message {
             match self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { callback_data } => Message {
                     sender_id: sender_id,
                     opcode: 0,
@@ -967,7 +933,6 @@ pub mod wl_callback {
             F: FnOnce(u32, &mut [wl_argument]) -> T,
         {
             match self {
-                Event::__nonexhaustive => unreachable!(),
                 Event::Done { callback_data } => {
                     let mut _args_array: [wl_argument; 1] = unsafe { ::std::mem::zeroed() };
                     _args_array[0].u = callback_data;