+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [client/server] Introduce the `DispatchData` mechanism, allowing global state to be shared with the
   filters and callbaks during an event dispatch.
+- [scanner] Use `#[non_exhaustive]` on protocol-generated enums to replace the `__nonexhaustive` hidden variant
 
 ## 0.24.1 -- 2019-12-13